capacity(local): restructure index.js exports for workerd compatibility
