+++ tests/security-headers.test.js
@@ -13,7 +13,7 @@ import {
   applySecurityHeaders,
   serialiseHeadersBlock,
 } from '../worker/src/security-headers.js';
-import { applySecurityHeadersSafely } from '../worker/src/index.js';
+import { applySecurityHeadersSafely } from '../worker/src/security-headers-safe.js';
 import { assertHeadersBlockIsFresh } from '../scripts/lib/headers-drift.mjs';
 import { createWorkerRepositoryServer } from './helpers/worker-server.js';
